Initial commit of the mob-parallel branch. This branche aims to create a fully parallelized version of Dos9, in order to create an experimental command prompt that superseeds cmd.exe and may rivalize with other shells.
The design perspective are *full* parallelization of the Dos9 core programs
to make it possible to run various batch scripts simultaneously  in the same
process so that they can share data between them and create somewhat a
concurrent programming language.

Works has already been started on the following points :

	- Change of the stream handling architectures, which remains a stack
though the streams are never dup2'ed until the interpretors really need to
run an external command.

	- A custom environment variable for Dos9, in order to allow different
threads to have dinstinct environment variables, although still permitting
sharing variables between threads. It also provides more compatibility between
various operating systems (for example use of "=" in variable names is not
allowed on GNU/Linux systems while it is on MS-Windows systems).

	- Contextification of the functions. This is the essential point that
ensure whole functions are thread safe.

	- Enhancement of FOR loop and CALL functions, according to new
directives.

	- New errors handling system. The purpose of Dos9_ShowErrorMessage
is to serve errors that will result in program's exit, thus, it is designed
to handle error message of the core functions on which an error will
eventually require process to be shut down.

If the parallel version of Dos9 is shown to be more robust, and quicker (and
I obviously hope it will), the parallel branch will be the main Dos9 branch.