Huge changeset modification for this commit,
These are still related to migration to a multithreaded model than to a
a multiprocessed model (which is common in *NIX world but can be *really*
hard to use in Microsoft Windows, with fork times significantly higher).

Basically, the source does not compile yet, there's still much work
to be done to make the source compilable. Whatever, I broke some little
functionnalities so that, even when we will be able to build the executable,
the resulting interpreter will have little restrictions (some broken for, at
least)

The status of the job is about to end contextualization of the source code,
and make it threads safe. Some parts will requiere more attentive attention,
especially the SET/A expression processing.